move functionality out of __init__.py
